requirements.txt.txt
flask
rembg
pillow
gunicorn